Backend/requirements.txt
flask
flask-cors
python-dotenv
bcrypt
PyJWT
pymongo
python-docx
Tesseract
pytesseract
poppler-utils
pdf2image
psycopg2
pinecone-client==2.2.2
openai==0.27.10
langchain==0.0.250
PyPDF2==3.0.1
python-dotenv==1.0.0
tiktoken